Created seperated hook to handle navigation after inactivity, created a seperated modal dedicated to display the sessionexpired message modal.
