+++ client/src/components/Modals/messageModal/MessageModal.module.css
@@ -5,19 +5,19 @@ dialog:not([open]) {
 .modal {
 	display: flex;
 	flex-direction: column;
-	padding-inline: 1rem;
-	padding-block: 0.5rem;
+	padding-inline: var(--spacing-4);
+	padding-block: var(--spacing-2);
 	margin-inline: auto;
 	margin-top: 25dvh;
-	border-radius: 0.25rem;
+	border-radius: var(--spacing-1);
 	border: 1px solid var(--primary-color);
 	& h2 {
 		color: var(--secondary-color);
-		padding-bottom: 1.25rem;
+		padding-bottom: var(--spacing-5);
 	}
 	& button {
-		margin-top: 2.5rem;
-		margin-bottom: 0.25rem;
+		margin-top: var(--spacing-10);
+		margin-bottom: var(--spacing-1);
 		margin-inline: auto;
 	}
 }
