+++ client/src/components/ui/button/Button.tsx
@@ -24,4 +24,4 @@ const Button = ({
 	);
 };
 
-export default Button;
+export { Button };
