+++ client/src/pages/homePage/HomePage.tsx
@@ -1,27 +1,9 @@
-import { useSearchParams } from "react-router-dom";
 import styles from "./HomePage.module.css";
-import { useEffect, useState } from "react";
-import { MessageModal } from "@components/MessageModal/MessageModal";
+import { SessionExpiredModal } from "@components/Modals/sessionExpiredModal/SessionExpiredModal";
 const HomePage = () => {
-	const [searchParams, setSearchParams] = useSearchParams();
-	const [showModal, setShowModal] = useState(searchParams.get("sessionExpired") === "true");
-
-	useEffect(() => {
-		if (searchParams.has("sessionExpired")) {
-			const newSearchParams = new URLSearchParams(searchParams);
-			newSearchParams.delete("sessionExpired");
-			setSearchParams(newSearchParams, { replace: true });
-		}
-	}, [searchParams, setSearchParams]);
-
 	return (
 		<>
-			<MessageModal
-				title="Inactivity detected"
-				message="You have been automatically sent to the home page because you ware inactive for 5 minutes."
-				isOpen={showModal}
-				closeModal={() => setShowModal(false)}
-			/>
+			<SessionExpiredModal />
 			<section className={styles.section}>
 				<h2>Home</h2>
 				<p>Welcome to Square Creator. This project is the result of my attempt to tackle Wizardworks code challenge.</p>
