+++ server/api/Data/LastSquareData.json
@@ -1,6 +1,6 @@
 {
-  "Id": 95,
-  "X": 5,
-  "Y": 9,
-  "Color": "#818997"
+  "Id": 16,
+  "X": 0,
+  "Y": 3,
+  "Color": "#aa3e88"
 }
