+++ server/api/Data/SquareData.json
@@ -3,570 +3,96 @@
     "Id": 1,
     "X": 0,
     "Y": 0,
-    "Color": "#35e477"
+    "Color": "#e60f59"
   },
   {
     "Id": 2,
     "X": 1,
     "Y": 0,
-    "Color": "#acfd85"
+    "Color": "#e2d9f7"
   },
   {
     "Id": 3,
     "X": 1,
     "Y": 1,
-    "Color": "#4feaf8"
+    "Color": "#037a7c"
   },
   {
     "Id": 4,
     "X": 0,
     "Y": 1,
-    "Color": "#9ec947"
+    "Color": "#3ad15a"
   },
   {
     "Id": 5,
     "X": 2,
     "Y": 0,
-    "Color": "#20a9cd"
+    "Color": "#8cfd21"
   },
   {
     "Id": 6,
     "X": 2,
     "Y": 1,
-    "Color": "#557dd5"
+    "Color": "#31c8bc"
   },
   {
     "Id": 7,
     "X": 2,
     "Y": 2,
-    "Color": "#ee4043"
+    "Color": "#d5fe7b"
   },
   {
     "Id": 8,
     "X": 1,
     "Y": 2,
-    "Color": "#32d40c"
+    "Color": "#e3198f"
   },
   {
     "Id": 9,
     "X": 0,
     "Y": 2,
-    "Color": "#a2c972"
+    "Color": "#639f08"
   },
   {
     "Id": 10,
     "X": 3,
     "Y": 0,
-    "Color": "#f28212"
+    "Color": "#c80929"
   },
   {
     "Id": 11,
     "X": 3,
     "Y": 1,
-    "Color": "#37ef7b"
+    "Color": "#2df186"
   },
   {
     "Id": 12,
     "X": 3,
     "Y": 2,
-    "Color": "#3209f1"
+    "Color": "#2aaf76"
   },
   {
     "Id": 13,
     "X": 3,
     "Y": 3,
-    "Color": "#8b551f"
+    "Color": "#331a6b"
   },
   {
     "Id": 14,
     "X": 2,
     "Y": 3,
-    "Color": "#38342b"
+    "Color": "#3d15e2"
   },
   {
     "Id": 15,
     "X": 1,
     "Y": 3,
-    "Color": "#3031d6"
+    "Color": "#052146"
   },
   {
     "Id": 16,
     "X": 0,
     "Y": 3,
-    "Color": "#84d19e"
-  },
-  {
-    "Id": 17,
-    "X": 4,
-    "Y": 0,
-    "Color": "#ccbaae"
-  },
-  {
-    "Id": 18,
-    "X": 4,
-    "Y": 1,
-    "Color": "#32ce16"
-  },
-  {
-    "Id": 19,
-    "X": 4,
-    "Y": 2,
-    "Color": "#339594"
-  },
-  {
-    "Id": 20,
-    "X": 4,
-    "Y": 3,
-    "Color": "#a7c321"
-  },
-  {
-    "Id": 21,
-    "X": 4,
-    "Y": 4,
-    "Color": "#386b76"
-  },
-  {
-    "Id": 22,
-    "X": 3,
-    "Y": 4,
-    "Color": "#d56b56"
-  },
-  {
-    "Id": 23,
-    "X": 2,
-    "Y": 4,
-    "Color": "#918ff4"
-  },
-  {
-    "Id": 24,
-    "X": 1,
-    "Y": 4,
-    "Color": "#cea587"
-  },
-  {
-    "Id": 25,
-    "X": 0,
-    "Y": 4,
-    "Color": "#575c5e"
-  },
-  {
-    "Id": 26,
-    "X": 5,
-    "Y": 0,
-    "Color": "#911b8b"
-  },
-  {
-    "Id": 27,
-    "X": 5,
-    "Y": 1,
-    "Color": "#873e43"
-  },
-  {
-    "Id": 28,
-    "X": 5,
-    "Y": 2,
-    "Color": "#ebb32e"
-  },
-  {
-    "Id": 29,
-    "X": 5,
-    "Y": 3,
-    "Color": "#dac5ee"
-  },
-  {
-    "Id": 30,
-    "X": 5,
-    "Y": 4,
-    "Color": "#217fd4"
-  },
-  {
-    "Id": 31,
-    "X": 5,
-    "Y": 5,
-    "Color": "#b577b7"
-  },
-  {
-    "Id": 32,
-    "X": 4,
-    "Y": 5,
-    "Color": "#85b1b8"
-  },
-  {
-    "Id": 33,
-    "X": 3,
-    "Y": 5,
-    "Color": "#20f833"
-  },
-  {
-    "Id": 34,
-    "X": 2,
-    "Y": 5,
-    "Color": "#48d238"
-  },
-  {
-    "Id": 35,
-    "X": 1,
-    "Y": 5,
-    "Color": "#549cc0"
-  },
-  {
-    "Id": 36,
-    "X": 0,
-    "Y": 5,
-    "Color": "#e58782"
-  },
-  {
-    "Id": 37,
-    "X": 6,
-    "Y": 0,
-    "Color": "#0f304b"
-  },
-  {
-    "Id": 38,
-    "X": 6,
-    "Y": 1,
-    "Color": "#80399b"
-  },
-  {
-    "Id": 39,
-    "X": 6,
-    "Y": 2,
-    "Color": "#e640bc"
-  },
-  {
-    "Id": 40,
-    "X": 6,
-    "Y": 3,
-    "Color": "#0b49ca"
-  },
-  {
-    "Id": 41,
-    "X": 6,
-    "Y": 4,
-    "Color": "#a8182c"
-  },
-  {
-    "Id": 42,
-    "X": 6,
-    "Y": 5,
-    "Color": "#fd4f50"
-  },
-  {
-    "Id": 43,
-    "X": 6,
-    "Y": 6,
-    "Color": "#f68d75"
-  },
-  {
-    "Id": 44,
-    "X": 5,
-    "Y": 6,
-    "Color": "#008ab0"
-  },
-  {
-    "Id": 45,
-    "X": 4,
-    "Y": 6,
-    "Color": "#f32bea"
-  },
-  {
-    "Id": 46,
-    "X": 3,
-    "Y": 6,
-    "Color": "#939693"
-  },
-  {
-    "Id": 47,
-    "X": 2,
-    "Y": 6,
-    "Color": "#421a82"
-  },
-  {
-    "Id": 48,
-    "X": 1,
-    "Y": 6,
-    "Color": "#fadc15"
-  },
-  {
-    "Id": 49,
-    "X": 0,
-    "Y": 6,
-    "Color": "#c7f6d2"
-  },
-  {
-    "Id": 50,
-    "X": 7,
-    "Y": 0,
-    "Color": "#11cd97"
-  },
-  {
-    "Id": 51,
-    "X": 7,
-    "Y": 1,
-    "Color": "#b9e418"
-  },
-  {
-    "Id": 52,
-    "X": 7,
-    "Y": 2,
-    "Color": "#935cca"
-  },
-  {
-    "Id": 53,
-    "X": 7,
-    "Y": 3,
-    "Color": "#28e9d3"
-  },
-  {
-    "Id": 54,
-    "X": 7,
-    "Y": 4,
-    "Color": "#702ae4"
-  },
-  {
-    "Id": 55,
-    "X": 7,
-    "Y": 5,
-    "Color": "#0e85a3"
-  },
-  {
-    "Id": 56,
-    "X": 7,
-    "Y": 6,
-    "Color": "#fe6fef"
-  },
-  {
-    "Id": 57,
-    "X": 7,
-    "Y": 7,
-    "Color": "#ff10e7"
-  },
-  {
-    "Id": 58,
-    "X": 6,
-    "Y": 7,
-    "Color": "#87c008"
-  },
-  {
-    "Id": 59,
-    "X": 5,
-    "Y": 7,
-    "Color": "#d61ead"
-  },
-  {
-    "Id": 60,
-    "X": 4,
-    "Y": 7,
-    "Color": "#24a715"
-  },
-  {
-    "Id": 61,
-    "X": 3,
-    "Y": 7,
-    "Color": "#da3693"
-  },
-  {
-    "Id": 62,
-    "X": 2,
-    "Y": 7,
-    "Color": "#0b57e6"
-  },
-  {
-    "Id": 63,
-    "X": 1,
-    "Y": 7,
-    "Color": "#bd5167"
-  },
-  {
-    "Id": 64,
-    "X": 0,
-    "Y": 7,
-    "Color": "#5a0647"
-  },
-  {
-    "Id": 65,
-    "X": 8,
-    "Y": 0,
-    "Color": "#50f07e"
-  },
-  {
-    "Id": 66,
-    "X": 8,
-    "Y": 1,
-    "Color": "#f4b91c"
-  },
-  {
-    "Id": 67,
-    "X": 8,
-    "Y": 2,
-    "Color": "#508d20"
-  },
-  {
-    "Id": 68,
-    "X": 8,
-    "Y": 3,
-    "Color": "#cc2139"
-  },
-  {
-    "Id": 69,
-    "X": 8,
-    "Y": 4,
-    "Color": "#813f55"
-  },
-  {
-    "Id": 70,
-    "X": 8,
-    "Y": 5,
-    "Color": "#f919dd"
-  },
-  {
-    "Id": 71,
-    "X": 8,
-    "Y": 6,
-    "Color": "#9d751a"
-  },
-  {
-    "Id": 72,
-    "X": 8,
-    "Y": 7,
-    "Color": "#1cccd9"
-  },
-  {
-    "Id": 73,
-    "X": 8,
-    "Y": 8,
-    "Color": "#6af4bc"
-  },
-  {
-    "Id": 74,
-    "X": 7,
-    "Y": 8,
-    "Color": "#4c041d"
-  },
-  {
-    "Id": 75,
-    "X": 6,
-    "Y": 8,
-    "Color": "#0bb470"
-  },
-  {
-    "Id": 76,
-    "X": 5,
-    "Y": 8,
-    "Color": "#21e4d8"
-  },
-  {
-    "Id": 77,
-    "X": 4,
-    "Y": 8,
-    "Color": "#1c6145"
-  },
-  {
-    "Id": 78,
-    "X": 3,
-    "Y": 8,
-    "Color": "#ab22dd"
-  },
-  {
-    "Id": 79,
-    "X": 2,
-    "Y": 8,
-    "Color": "#106ee0"
-  },
-  {
-    "Id": 80,
-    "X": 1,
-    "Y": 8,
-    "Color": "#1a0a47"
-  },
-  {
-    "Id": 81,
-    "X": 0,
-    "Y": 8,
-    "Color": "#3ced6b"
-  },
-  {
-    "Id": 82,
-    "X": 9,
-    "Y": 0,
-    "Color": "#97d618"
-  },
-  {
-    "Id": 83,
-    "X": 9,
-    "Y": 1,
-    "Color": "#8f49bc"
-  },
-  {
-    "Id": 84,
-    "X": 9,
-    "Y": 2,
-    "Color": "#fbbc15"
-  },
-  {
-    "Id": 85,
-    "X": 9,
-    "Y": 3,
-    "Color": "#07eaa4"
-  },
-  {
-    "Id": 86,
-    "X": 9,
-    "Y": 4,
-    "Color": "#666cde"
-  },
-  {
-    "Id": 87,
-    "X": 9,
-    "Y": 5,
-    "Color": "#9213d5"
-  },
-  {
-    "Id": 88,
-    "X": 9,
-    "Y": 6,
-    "Color": "#b629ef"
-  },
-  {
-    "Id": 89,
-    "X": 9,
-    "Y": 7,
-    "Color": "#5b8673"
-  },
-  {
-    "Id": 90,
-    "X": 9,
-    "Y": 8,
-    "Color": "#ad739d"
-  },
-  {
-    "Id": 91,
-    "X": 9,
-    "Y": 9,
-    "Color": "#fc8c05"
-  },
-  {
-    "Id": 92,
-    "X": 8,
-    "Y": 9,
-    "Color": "#f3b484"
-  },
-  {
-    "Id": 93,
-    "X": 7,
-    "Y": 9,
-    "Color": "#e1de82"
-  },
-  {
-    "Id": 94,
-    "X": 6,
-    "Y": 9,
-    "Color": "#775125"
-  },
-  {
-    "Id": 95,
-    "X": 5,
-    "Y": 9,
-    "Color": "#818997"
+    "Color": "#aa3e88"
   }
 ]
